+++ CHANGELOG.md
@@ -44,6 +44,9 @@
 - Enh #203: Add `getGuestRoleName()` and `getGuestRole()` methods to `Manager` (@arogachev)
 - Chg #203: Throw `RuntimeException` in the case with implicit guest and non-existing guest role in 
 `Manager::userHasPermission()` (@arogachev)
+- Enh #206: Optimize calls for getting child items within the loops (@arogachev)
+- Chg #206: Rename `$name` argument to `$names` and allow array type for it in `getAllChildren()`, `getAllChildRoles()`,
+  `getAllChildPermissions()` methods of `ItemsStorageInterface` (@arogachev)
 
 ## 1.0.2 April 20, 2023